Patching up some logic errors

+++ commands/rss-feed.js
@@ -1,4 +1,4 @@
-import { SlashCommandBuilder, EmbedBuilder, inlineCode, TextChannel } from 'discord.js';
+import { SlashCommandBuilder, EmbedBuilder, inlineCode, ChannelType } from 'discord.js';
 import RSSObj from '../models/RSSObj.js';
 import logclass from '../logger.js';
 import { get_latest } from '../submodules/feed-parser.js';
@@ -82,12 +82,10 @@ export async function handler(interaction) {
           // webhook cleanup
           const exists = await RSSObj.find({ channel_id: callback.channel_id }).exec();
           const channel = await bot_client.channels.cache.get(callback.channel_id).fetch();
-          if (exists.length < 1 && channel == TextChannel.prototype) {
+          if (exists.length < 1 && channel.type == ChannelType.GuildText) {
             const webhooks = await channel.fetchWebhooks();
             const rss = webhooks.find(hook => hook.name === `${bot_client.user.tag} - RSS`);
-            if (rss != undefined) {
-              rss.delete();
-            }
+            if (rss != undefined) { rss.delete(); }
           }
         }
       });
